requirements.txt
torch>=2.0.0
torchvision>=0.15.0
clip-by-openai>=0.2.0
faiss-cpu>=1.7.4
streamlit>=1.27.0
Pillow>=9.5.0
numpy>=1.24.0